GH-121 Abstract metadata storage with MetadataBackend trait (Phase 3)
Introduce a MetadataBackend trait that abstracts version-control metadata
operations (commits, branches, history, file access) away from the
concrete gix::Repository dependency.

Key changes:
- New src/git/metadata.rs: MetadataBackend trait (16 methods) and
  GitMetadataBackend implementation wrapping gix::Repository
- VersionedKvStore gains a third type parameter M: MetadataBackend
  with default M = GitMetadataBackend (all type aliases unchanged)
- core.rs and history.rs now use trait methods instead of direct
  gix::Repository access, reducing git-specific code by ~600 lines
- backends.rs retains direct repo access via GitMetadataBackend::repo()
  for storage-specific operations (GitNodeStorage, operations.rs)
- git_repo() backward-compat method preserved on GitVersionedKvStore

This makes VersionedKvStore pluggable over metadata backends. A future
InMemoryMetadataBackend or FileMetadataBackend could be used without
any git dependency for testing or embedded use cases.
